+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - feat(template.root): Add link to list of Incident/Maintenance (#11) (!53)
 - feat(template.scope): Add link to list of Incident/Maintenance (#11) (!53)
 - feat(template.service): Add link to list of Incident/Maintenance (#11) (!53)
+- feat(list.template): Display current resources list filters (#13) (!55)
 
 ### Fixed